creating new endpoints and changing the structure of views
