+++ django_project_491/django_app/urls.py
@@ -1,21 +1,41 @@
 from django.urls import path
-from .views import *
+from .views import comment_views, question_views, user_views, utilization_views
+
 from django.contrib.auth.views import LoginView, LogoutView
 
 urlpatterns = [
-    path('wikipedia/', wikipedia_data_views, name='wikipedia'),
-    path('login/', login_user, name='login'),
-    path('signup/', signup, name='signup'),
-    path('home/', home, name='home'),
+
+    path('home/', utilization_views.home, name='home'),
+
+    path('wikipedia/', utilization_views.wikipedia_data_views, name='wikipedia'),
+    path('search/<str:search_strings>', utilization_views.wiki_search, name='wiki_search'),
+    path('result/<str:wiki_id>', utilization_views.wiki_result, name='wiki_result'),
+    path('get_api_languages/', utilization_views.get_run_coder_api_languages, name='get_run_coder_api_languages'),
+
+    path('login/', user_views.login_user, name='login'),
+    path('signup/', user_views.signup, name='signup'),
     path('logout/', LogoutView.as_view(), name='logout'),
-    path('search/<str:search_strings>', wiki_search, name='wiki_search'),
-    path('result/<str:wiki_id>', wiki_result, name='wiki_result'),
-    path('create_comment/', create_comment, name='create_comment'),
-    path('create_question/', create_question, name='create_question'),
-    path('list_questions/', list_questions_by_language, name='list_questions'),
-    path('run_code/', run_code_view, name='run_code'),
-    path('get_api_languages/', get_run_coder_api_languages, name='get_run_coder_api_languages'),
-    path('random_questions/', random_questions, name='random_questions'),
-    path('question/<int:question_id>/comments/', get_question_comments, name='get_question_comments'),
-    path('code_execute/', post_sample_code, name='code_execute'), # for dynamic code execution
+
+    path('create_question/', question_views.create_question, name='create_question'),
+    path('get_question/<int:question_id>/', question_views.get_question, name='get_question'),
+    path('get_question/<int:question_id>/', question_views.get_question, name='get_question'),
+    path('edit_question/<int:question_id>/', question_views.edit_question, name='edit_question'),
+    path('delete_question/<int:question_id>/', question_views.delete_question, name='delete_question'),
+    path('mark_as_answered/<int:question_id>/', question_views.mark_as_answered, name='mark_as_answered'), # TODO: IMPLEMENT THE LOGIC
+    path('report_question/<int:question_id>/', question_views.report_question, name='report_question'),
+    path('get_questions_by_hotness/<int:page_number>', question_views.list_questions_by_hotness, name='get_question_comments'),
+    path('question/<int:question_id>/comments/', question_views.get_question_comments, name='get_question_comments'),
+    path('random_questions/', question_views.random_questions, name='random_questions'),
+    path('list_questions/', question_views.list_questions_by_language, name='list_questions'),
+
+    path('upvote_object/<int:question_id>/', utilization_views.upvote_object, name='upvote_object'), # TODO: CHANGE THE UPVOTE LOGIC
+    path('downvote_object/<int:question_id>/', utilization_views.downvote_object, name='downvote_object'), # TODO: CHANGE THE DOWNVOTE LOGIC
+
+    path('create_comment/', comment_views.create_comment, name='create_comment'),
+    path('edit_comment/<int:comment_id>', comment_views.edit_comment, name='edit_comment'),
+    path('delete_comment/<int:comment_id>', comment_views.delete_comment, name='delete_comment'),
+
+    path('code_execute/', utilization_views.post_sample_code, name='code_execute'), # for dynamic code execution
+    path('run_code/', utilization_views.run_code_view, name='run_code'),
+
 ]
